fix: Correcciones sonar

+++ src/app/routes/cars/brand/one/brand-one.page.ts
@@ -62,12 +62,8 @@ export class BrandOnePage implements OnInit {
     }
 
     actionForOption(option: ActionForOptionI) {
-        switch (option.value) {
-            case 'delete':
-                this.delete();
-                break;
-            default:
-                break;
+        if (option.value === 'delete') {
+            this.delete();
         }
     }
