+++ src/app/routes/system/toggles/pages/one/toggles-one.page.ts
@@ -42,12 +42,8 @@ export class TogglesOnePage {
     }
 
     actionForOption(option: ActionForOptionI) {
-        switch (option.value) {
-            case 'delete':
-                this.delete();
-                break;
-            default:
-                break;
+        if (option.value === 'delete') {
+            this.delete();
         }
     }
